requirements.txt
bs4
dash
numpy
Flask
plotly
pandas
requests
gunicorn
pandas-ta
flask-restful 
python-dotenv
python-binance
python-telegram-bot
